integrations/mcp/requirements.txt
hyperliquid-python-sdk>=0.17.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
aiohttp>=3.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
eth-account>=0.10.0